+++ src/ggrc-client/js/plugins/tests/import-export-utils_spec.js
@@ -3,7 +3,7 @@
   Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
 
-import * as importExportUtils from '../import-export-utils';
+import * as importExportUtils from '../utils/import-export-utils';
 
 const baseUrlPrefix = '/api/people/1';
